+++ SNetwork Sensor/src/util/Protocol.java
@@ -5,8 +5,8 @@ package util;
  * @author Matheus Teles
  */
 public final class Protocol {
-	public static final int HI = 0;
-	public static final int INFO = 1;
-	public static final int DATA = 2;
-	public static final int BYE = 3;
+	public static final byte HI = 0;
+	public static final byte INFO = 1;
+	public static final byte DATA = 2;
+	public static final byte BYE = 3;
 }
